requirements.txt
# Core ML and Computer Vision
tensorflow>=2.10.0
keras-facenet>=0.3.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
Pillow>=9.0.0
rawpy>=0.18.0

# Web Framework
Flask>=2.2.0
Flask-CORS>=3.0.10

# Data Processing
numpy>=1.21.0
scipy>=1.9.0
h5py>=3.7.0

# Utilities
requests>=2.28.0
psutil>=5.9.0

# Development Tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Performance (optional)
gunicorn>=20.1.0

